Store active tab before opening survey in new tab (closes #175).
